requirements.txt
numpy==1.26.4
openai==1.40.3
openpyxl==3.1.5
pandasai==2.2.14
python-dotenv==1.0.1
streamlit==1.37.1
scikit-learn==1.5.1
PyYAML=6.0.2